+++ obj/src/connect/index.js
@@ -1,10 +1,11 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.MongoDbConnectionResolver = void 0;
 /**
  * @module connect
  * @preferred
  *
  */
 var MongoDbConnectionResolver_1 = require("./MongoDbConnectionResolver");
-exports.MongoDbConnectionResolver = MongoDbConnectionResolver_1.MongoDbConnectionResolver;
+Object.defineProperty(exports, "MongoDbConnectionResolver", { enumerable: true, get: function () { return MongoDbConnectionResolver_1.MongoDbConnectionResolver; } });
 //# sourceMappingURL=index.js.map