+++ obj/src/persistence/MongoDbIndex.js
@@ -1,6 +1,8 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.MongoDbIndex = void 0;
 /**
+ * @module persistence
  * Index definition for mondodb
  */
 class MongoDbIndex {