+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> MongoDB components for Node.js Changelog
 
+## <a name="3.5.0"></a> 3.5.0 (2021-02-18) 
+
+### Features
+* Added defineSchema method that shall be overriden in child classes
+* Added clearSchema method
+
 ## <a name="3.4.0"></a> 3.4.0 (2020-06-09) 
 
 ### Features